requirements/requirements.txt
wheel
django-anymail==8.4
django-allauth==0.45.0
django-compressor==2.4
django-sass-processor==0.8.*
Django>=3.2.2,<3.3
libsass==0.20.*
psycopg2==2.9.1
python-dotenv==0.19.0
wagtail-metadata==3.4.0
wagtail>=2.14,<2.15
wagtailmenus>=3.0
